added CDF scaling in options
In the validate page, an option has been added to the dropdown menu to use the recently implemented CDF scaling function (cdf_beta_match) of pytesmo. The environment file has been updated to link to the git page of pytesmo for the qa4sm check. After the new pytesmo release, this should be changed to point at the new version.