Cdf matching issue 309 (#151)
* Added quick inspection table in results

An (expandable) table was added under the section 'Summary Statistics' in the results page.

* Improved table description in results page

* Fixed bug in validation page

A problem occurred as the inspection table function was called before the output results were ready, leading to an error in the validation run page. Small typo fixed in results table

* Added measure units to results; added CDF matching option

The summary statistics table now includes the measure units for each metric. The CDF scaling option has been included in the sclaing section of the validate page

* Updated .yml with git reader version and amended pytesmo CDF implementation

* fixed issue in tests

* added CDF scaling in options

In the validate page, an option has been added to the dropdown menu to use the recently implemented CDF scaling function (cdf_beta_match) of pytesmo. The environment file has been updated to link to the git page of pytesmo for the qa4sm check. After the new pytesmo release, this should be changed to point at the new version.

* updated tests

The (deprecated) pytesmo matching functions have been removed from the code; and the tests have been updated to run a validation with the new CDF beta matching sclaing function

* restored files

Changes belonging to previous commitments have been restored to only have the CDF_matching_issue-309

* updated results.py

* Update qa4sm_env.yml

* Update qa4sm_env.yml

Co-authored-by: Wolfgang Preimesberger <wolfgang.preimesberger@geo.tuwien.ac.at>